refactor: add custom hook for handling "likes" to reduce code duplication
